perf(inbound): trim reply-run startup imports
