+++ src/agents/pi-embedded.runtime.ts
@@ -1,0 +1,6 @@
+export {
+  abortEmbeddedPiRun,
+  isEmbeddedPiRunActive,
+  isEmbeddedPiRunStreaming,
+  resolveEmbeddedSessionLane,
+} from "./pi-embedded.js";
