+++ src/auto-reply/reply/session-system-events.ts
@@ -1,0 +1,105 @@
+import { resolveUserTimezone } from "../../agents/date-time.js";
+import type { OpenClawConfig } from "../../config/config.js";
+import { buildChannelSummary } from "../../infra/channel-summary.js";
+import {
+  formatUtcTimestamp,
+  formatZonedTimestamp,
+  resolveTimezone,
+} from "../../infra/format-time/format-datetime.ts";
+import { drainSystemEventEntries } from "../../infra/system-events.js";
+
+/** Drain queued system events, format as `System:` lines, return the block (or undefined). */
+export async function drainFormattedSystemEvents(params: {
+  cfg: OpenClawConfig;
+  sessionKey: string;
+  isMainSession: boolean;
+  isNewSession: boolean;
+}): Promise<string | undefined> {
+  const compactSystemEvent = (line: string): string | null => {
+    const trimmed = line.trim();
+    if (!trimmed) {
+      return null;
+    }
+    const lower = trimmed.toLowerCase();
+    if (lower.includes("reason periodic")) {
+      return null;
+    }
+    if (lower.startsWith("read heartbeat.md")) {
+      return null;
+    }
+    if (lower.includes("heartbeat poll") || lower.includes("heartbeat wake")) {
+      return null;
+    }
+    if (trimmed.startsWith("Node:")) {
+      return trimmed.replace(/ · last input [^·]+/i, "").trim();
+    }
+    return trimmed;
+  };
+
+  const resolveSystemEventTimezone = (cfg: OpenClawConfig) => {
+    const raw = cfg.agents?.defaults?.envelopeTimezone?.trim();
+    if (!raw) {
+      return { mode: "local" as const };
+    }
+    const lowered = raw.toLowerCase();
+    if (lowered === "utc" || lowered === "gmt") {
+      return { mode: "utc" as const };
+    }
+    if (lowered === "local" || lowered === "host") {
+      return { mode: "local" as const };
+    }
+    if (lowered === "user") {
+      return {
+        mode: "iana" as const,
+        timeZone: resolveUserTimezone(cfg.agents?.defaults?.userTimezone),
+      };
+    }
+    const explicit = resolveTimezone(raw);
+    return explicit ? { mode: "iana" as const, timeZone: explicit } : { mode: "local" as const };
+  };
+
+  const formatSystemEventTimestamp = (ts: number, cfg: OpenClawConfig) => {
+    const date = new Date(ts);
+    if (Number.isNaN(date.getTime())) {
+      return "unknown-time";
+    }
+    const zone = resolveSystemEventTimezone(cfg);
+    if (zone.mode === "utc") {
+      return formatUtcTimestamp(date, { displaySeconds: true });
+    }
+    if (zone.mode === "local") {
+      return formatZonedTimestamp(date, { displaySeconds: true }) ?? "unknown-time";
+    }
+    return (
+      formatZonedTimestamp(date, { timeZone: zone.timeZone, displaySeconds: true }) ??
+      "unknown-time"
+    );
+  };
+
+  const systemLines: string[] = [];
+  const queued = drainSystemEventEntries(params.sessionKey);
+  systemLines.push(
+    ...queued
+      .map((event) => {
+        const compacted = compactSystemEvent(event.text);
+        if (!compacted) {
+          return null;
+        }
+        return `[${formatSystemEventTimestamp(event.ts, params.cfg)}] ${compacted}`;
+      })
+      .filter((v): v is string => Boolean(v)),
+  );
+  if (params.isMainSession && params.isNewSession) {
+    const summary = await buildChannelSummary(params.cfg);
+    if (summary.length > 0) {
+      systemLines.unshift(...summary);
+    }
+  }
+  if (systemLines.length === 0) {
+    return undefined;
+  }
+
+  return systemLines
+    .flatMap((line) => line.split("\n").map((subline) => `System: ${subline}`))
+    .join("\n");
+}
